move component into it's corresponding files
